# Optimization: remove `updateClientMemUsage` from i/o threads.