--- a/src/eval.c
+++ b/src/eval.c
@@ -812,6 +812,7 @@ void ldbEndSession(client *c) {
     /* If it's a fork()ed session, we just exit. */
     if (ldb.forked) {
         writeToClient(c,0);
+        updateClientMemUsage(c);
         serverLog(LL_WARNING,"Lua debugging session child exiting");
         exitFromChild(0);
     } else {