@filipecosta90 2022-03-14
@yoav-steinberg we can notice a reduction of the CPU cycles on updateClientMemUsage from 2.2% -> 1.5% but the increase in ops/sec is not very significant ( as expected the measurements of throughput and latency are done without the profiler ). Nonetheless, if we look at the impact in latency ( 0.5% drop we can associate it with the drop on the CPU cycles spent on updateClientMemUsage ). 

WRT to measured changes on throughput and latency:
 - throughput: from 558591 to 559258 ops/sec ( 0.12% change )
 - latency (p50 including RTT): from 2.863 ms to 2.847 ms ( 0.5% change )

This is small improvement but nonetheless is a reduction that we should merge IMHO. Agree?


## confirmation that indeed the function is less heavy on CPU now

```
# to record data
perf record -g --pid `pgrep redis-server` --call-graph dwarf -o unstable -- sleep 30

# report ( search for updateClientMemUsage with `/` when in perf report mode )
perf report -g -g "graph,0.01,caller" -i unstable -d redis-server --inline
```

unstable (af6d5c5932fa4fc1a0461af5e3df50e9c284c4f9) total CPU time of updateClientMemUsage + children's:  ~2.2% of CPU cyles

```
Samples: 119K of event 'cycles:ppp', Event count (approx.): 107447366785
  Children      Self  Command       Symbol
+    2.17%     0.78%  redis-server  [.] updateClientMemUsage
+    0.23%     0.23%  redis-server  [.] updateClientMemUsageBucket
```

wip_optimize_updateClientMemUsage, (406aea42ac62866905805eadbf81e81cd2fa8b2e):

```
Samples: 119K of event 'cycles:ppp', Event count (approx.): 107509389716
  Children      Self  Command       Symbol
+    1.46%     0.31%  redis-server  [.] updateClientMemUsage
```

## benchmark 

steps:
```
# spin redis
taskset -c 0 ./src/redis-server --save "" --daemonize yes

# populate data
memtier_benchmark --ratio 1:0 --key-maximum 1000000 --key-minimum 1 --key-pattern P:P -d 1000 --hide-histogram

# benchmark
taskset -c 1,2 memtier_benchmark -d 1000 --ratio 0:1 --test-time 60 --pipeline 15 --key-pattern=P:P -t 2 --hide-histogram --key-maximum=1000000 --key-minimum 1 -x 3
```

results for unstable unstable (af6d5c5932fa4fc1a0461af5e3df50e9c284c4f9) :
```
BEST RUN RESULTS
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets            0.00          ---          ---             ---             ---             ---             ---         0.00 
Gets       558591.33    558591.33         0.00         2.68485         2.86300         3.82300         4.12700    568894.72 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     558591.33    558591.33         0.00         2.68485         2.86300         3.82300         4.12700    568894.72 


WORST RUN RESULTS
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets            0.00          ---          ---             ---             ---             ---             ---         0.00 
Gets       558335.05    558335.05         0.00         2.68610         2.84700         3.82300         4.12700    568634.38 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     558335.05    558335.05         0.00         2.68610         2.84700         3.82300         4.12700    568634.38 
```

results for wip_optimize_updateClientMemUsage, (406aea42ac62866905805eadbf81e81cd2fa8b2e) :
```
BEST RUN RESULTS
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets            0.00          ---          ---             ---             ---             ---             ---         0.00 
Gets       559258.27    559258.27         0.00         2.68165         2.84700         3.82300         4.12700    569574.49 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     559258.27    559258.27         0.00         2.68165         2.84700         3.82300         4.12700    569574.49 


WORST RUN RESULTS
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets            0.00          ---          ---             ---             ---             ---             ---         0.00 
Gets       558619.75    558619.75         0.00         2.68470         2.84700         3.82300         4.12700    568923.76 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     558619.75    558619.75         0.00         2.68470         2.84700         3.82300         4.12700    568923.76 

```