@yoav-steinberg 2022-03-14
@filipecosta90 Thanks for the analysis. Please note that this fix merges `updateClientMemUsage()` and `updateClientMemUsageBucket()` into one function. So the actual change in your benchmark from ~2.4% to ~1.5%.